+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## UNRELEASED
+
+### Added
+
+* Support for `AllowedPattern` in the `Environment` template parameter. (Issue #16)
+
+* Support for parameter files to have environment names that resemble regular expressions. If the parameter
+  file doesn't have a key that matches the environment name exactly, then each of the top-level keys in the
+  file will be interpreted as a regular expression and matched against the environment being deployed. The
+  first match wins, so if multiple keys can match, then the result is undefined. (Issue #16)
+
 ## [0.1.2] - 2017-12-22
 
 ### Changed